python/pyproject.toml
[tool.poetry]
name = "genart"
version = "0.1.0"
description = "Creative coding playground"
authors = ["Oliver Hofkens <oli.hofkens@gmail.com>"]
packages = [
    { include = "genart" }
]

[tool.poetry.scripts]
genart = 'genart.main:main'

[tool.poetry.dependencies]
python = "^3.9"

pycairo = "^1.21"
numpy = "^1.22"

black = { version = "^22.3", optional = true }
flake8 = { version = "^3.8", optional = true }
isort = { version = "^5.10", optional = true }
mypy = { version = "^0.950", optional = true }
pytest = { version = "^7.1", optional = true }
pytest-benchmark = { version = "^3.4", optional = true }
pytest-cov = { version = "*", optional = true }

[tool.poetry.extras]
linting = ["flake8", "black", "isort", "mypy"]
testing = ["pytest", "pytest-benchmark", "pytest-cov"]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
